rename table_day_generator.dart to day.dart
